Create README.md
Introduction
In predicting the rating of Apps in playstore there are many criterias which we have go through inorder to predict the best model for this data. Our dataset consists of some of the main criterias that are valid for the prediction of rating. Our dataset includes following variables that are taken in to consideration, Apps, Category, Reviws, Size, Installs, Type, Price, Content Rating , LastUpdated ,Genres, Current version, Android version.